backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
python-multipart>=0.0.18

# YAML parsing
pyyaml>=6.0.2

# Data validation
pydantic>=2.10.0

# NLP - Embeddings
sentence-transformers>=3.4.0

# NLP - Classifier
scikit-learn>=1.6.0
numpy>=1.26.0

# Claude API for data augmentation
anthropic>=0.42.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
httpx>=0.28.0
